requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.2.0
boto3==1.34.2
botocore==1.34.2
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.8
django-allauth==0.59.0
django-celery-beat==2.5.0
django-storages==1.14.2
django-timezone-field==6.1.0
django-widget-tweaks==1.5.0
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
kombu==5.3.4
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.9.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.12
